Documentation and Final Review
Final review of the v3.2 code, now live on our production server.

The documentation has been updated to reflect the work done during the summer of 2020 by Cannon Caspar as the author or most recent updater of each file.

Although not actively contributing during 2020, we must recognize Jeremiah DeGreeff who’s work has been the basis of or inspiration for the work that follows.

Co-Authored-By: Cannon Caspar <52666167+cannon544@users.noreply.github.com>