Basic Testing Table
Created a basic table displaying the testing data. Also (IMPORTANT) changed the name of one of the fields in the healthtest table. If already the earlier commit was loaded and installed, change the name of 'missed_test' in local_mxschool_healthtest to 'attended'. If not, then everything should be fine.