Create Code_of_Conduct.md
Here is my first draft of the code of conduct. It includes a small summary of the code, the rules, and a way to report violations and ask questions if necessary. Please edit or fix anything you see fit. I have left the text in one line because I assume when we create the doc, it will adjust like Chuck's document. If you know it does not do that, please feel free to change it.