Code of Conduct Statement (#72)
* Create Code_of_Conduct.md

Here is my first draft of the code of conduct. It includes a small summary of the code, the rules, and a way to report violations and ask questions if necessary. Please edit or fix anything you see fit. I have left the text in one line because I assume when we create the doc, it will adjust like Chuck's document. If you know it does not do that, please feel free to change it.

* Update and rename Code_of_Conduct.md to CODE_OF_CONDUCT.md

This is a strong start. I like the way you state our community standards concisely and directly. Should we also refer to the Student Handbook?

* Update CONTRIBUTING.md

Add link for the Code of Conduct document.

* Update CODE_OF_CONDUCT.md

Added a reference to the MX Student Handbook.
Tightened up the language a bit in an attempt to be more elegant.

Co-authored-by: CurryBoy9000 <55769093+CurryBoy9000@users.noreply.github.com>