Deans perm2.0 (#79)
* Deans Permission database update

With every big enhancement comes an annoying but necessary reconfiguration of the database table. Implementing this push will wipe all existing Dean's Permission Form data

* basic report layout

The javascript is scuffed but the report is all there.

* Comment javascript √

This one was tricky. Had to restructure the comment classes to be able to handle multiple comment fields in the same row. The javascript generally needs some major clean up

* small thing

* Status buttons

Another tricky one. Had to create an entirely new php, mustache, and javascript class. A few small bug fixes as well. The hardest parts are definitely done now--the rest should be fairly trivial.

* Working form

Form works. Things are going swimmingly

* notifications and preference form

Lots of tedious stuff here. Redid all of the notification classes and the preference form

* small bug fixes

* Notifications (needs testing)

For some reason emails aren't working on my local serve, so these emails need testing. Should send when a form is approved or denied, review is requested, when a form is submitted, and when the "email healthcenter" button is clicked.

* Healthcenter email, more sorting options, and bug fixes

* bug fix

* another bug fix?

* small thign

* email fix

This should work. Also there was an issue with the tag 'event date', which was previously just showing the unix time stamp. Now shows month/day/year

* 'other' event type no longer hardcoded

Co-authored-by: Cannon <cpcaspar@mxschool.edu>